+++ mex/mex.py
@@ -10,6 +10,8 @@ from mex.common.logging import logger
 from mex.editor.api.main import check_system_status, get_prometheus_metrics
 from mex.editor.aux_search.main import index as aux_search_index
 from mex.editor.aux_search.state import AuxState
+from mex.editor.create.main import index as create_index
+from mex.editor.create.state import CreateState
 from mex.editor.edit.main import index as edit_index
 from mex.editor.edit.state import EditState
 from mex.editor.login.main import index as login_index
@@ -56,6 +58,16 @@ app.add_page(
         State.load_nav,
         AuxState.refresh,
         AuxState.resolve_identifiers,
+    ],
+)
+app.add_page(
+    create_index,
+    route="/create",
+    title="MEx Editor | Create",
+    on_load=[
+        State.check_login,
+        State.load_nav,
+        CreateState.refresh,
     ],
 )
 app.add_page(